some tidy up
