--- octet/src/examples/MyMess/Test/Test.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Express 2013 for Windows Desktop
-VisualStudioVersion = 12.0.40629.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Test", "Test.vcxproj", "{3D30A6FA-EF22-4711-8733-A3B0CDC4A58F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3D30A6FA-EF22-4711-8733-A3B0CDC4A58F}.Debug|Win32.ActiveCfg = Debug|Win32
-		{3D30A6FA-EF22-4711-8733-A3B0CDC4A58F}.Debug|Win32.Build.0 = Debug|Win32
-		{3D30A6FA-EF22-4711-8733-A3B0CDC4A58F}.Release|Win32.ActiveCfg = Release|Win32
-		{3D30A6FA-EF22-4711-8733-A3B0CDC4A58F}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
