--- octet/src/examples/MyMess/Test/Test.vcxproj.filters
@@ -1,45 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Source Files">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="platform">
-      <UniqueIdentifier>{dda91860-e541-4fdb-a790-f6b5e7902ffb}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="scene">
-      <UniqueIdentifier>{1280c880-8181-435f-8975-ff6ac07df6ad}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="resources">
-      <UniqueIdentifier>{f85a3f01-4932-410d-b0e9-3861cb4ebf0d}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="loaders">
-      <UniqueIdentifier>{c05a7416-e0b3-4d3b-a560-c57b346f0665}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="containers">
-      <UniqueIdentifier>{579c6044-879b-4582-8dc0-08b19304347c}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="helpers">
-      <UniqueIdentifier>{294d83db-d00d-4c27-b636-2b796ecfd48b}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="math">
-      <UniqueIdentifier>{7c4ee1aa-1f06-43ef-9adf-8e1befbd9d0e}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="shaders">
-      <UniqueIdentifier>{22786083-47af-48b2-98c4-2963f667bc44}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="main.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="Test.h" />
-  </ItemGroup>
-</Project>
