+++ octet/src/examples/_PhysicsDemo/_PhysicsDemo.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Express 2013 for Windows Desktop
 VisualStudioVersion = 12.0.30723.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "example_shapes", "example_shapes.vcxproj", "{6722CC8F-3FC9-4B11-B7AE-918054DD5ACA}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "_PhysicsDemo", "_PhysicsDemo.vcxproj", "{6722CC8F-3FC9-4B11-B7AE-918054DD5ACA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
