+++ octet/src/examples/_PhysicsDemo/_PhysicsDemo.vcxproj.filters
@@ -427,7 +427,6 @@
     <ClInclude Include="..\..\containers\string.h">
       <Filter>containers</Filter>
     </ClInclude>
-    <ClInclude Include="example_shapes.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\resources\mesh_builder.inl">
