+++ octet/src/examples/_PhysicsDemo/_PhysicsDemo.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:example_shapes.xcodeproj">
+      location = "self:_PhysicsDemo.xcodeproj">
    </FileRef>
 </Workspace>
